This should address parts of both bug #1301353 and bug #1304770.
Instead of requiring a 'mongod' to be in PATH, we'll also search for
/usr/lib/juju/bin/mongod which is from the juju-mongodb package.
For core tests, we prefer the juju-mongodb version (since jujud itself
will prefer it), and for store tests we prefer 'mongod' on PATH since
it will have the v8 engine and allow us to run all the tests.

This also auto-detects when noTestMongoJS should be set because it is
using the /usr/lib/juju/bin/mongod executable. Note that this
autodetection *fails* for people who have been hacking around support by
symlinking /usr/lib/juju/bin/mongod back into their PATH. So I didn't
remove the ENV variable, but it should no longer be necessary.

I'm targetting 1.18 because with this patch I think Juju CI will have to
do less hackery to have the test suite JustWork for them.

https://codereview.appspot.com/97240043/

+++ store/mgo_test.go
@@ -37,16 +37,33 @@ func (s *MgoSuite) SetUpSuite(c *gc.C) {
 		"--smallfiles",
 		"--nojournal",
 	}
-	s.server = exec.Command("mongod", args...)
+	// Look for a system mongod first, since the juju-specific mongodb
+	// doesn't come with V8 so we have to skip some tests
+	mongopath, err := exec.LookPath("mongod")
+	if err == nil {
+		c.Log("found mongodb at: %q", mongopath)
+	} else {
+		c.Log("failed to find mongod: %v", err)
+		// We can fall back to /usr/lib/juju/bin/mongod but we should
+		// disable the JS tests
+		mongopath, err = exec.LookPath("/usr/lib/juju/bin/mongod")
+		c.Log("Tests requiring MongoDB Javascript will be skipped")
+		*noTestMongoJs = true
+	}
+	c.Assert(err, gc.IsNil)
+	s.server = exec.Command(mongopath, args...)
 	s.server.Stdout = &s.output
 	s.server.Stderr = &s.output
-	err := s.server.Start()
+	err = s.server.Start()
 	c.Assert(err, gc.IsNil)
 }
 
 func (s *MgoSuite) TearDownSuite(c *gc.C) {
-	s.server.Process.Kill()
-	s.server.Process.Wait()
+	if s.server != nil && s.server.Process != nil {
+		s.server.Process.Kill()
+		s.server.Process.Wait()
+		s.server.Process = nil
+	}
 }
 
 func (s *MgoSuite) SetUpTest(c *gc.C) {
